Update README.md
TMS FNC Gantt Chart v1.2.2.8 (March 6, 2025) Full Source

With the TMS FNC Gantt Chart you will get the project management tool to provide you with a visual of what needs to be done and when. The Gantt Chart is a component for tracking your project schedules. Where you can define if tasks rely on others, the progress that is made and the current status of your tasks.

[Download TMS FNC Gantt Chart v1.2.2.8 (March 6, 2025) Full Source](https://developer.team/delphi/35414-tms-fnc-gantt-chart-v1228-march-6-2025-full-source.html)